Interrupt working
Fixed Kernel and linker code to include call to constructor pointers
Fixed Port8BitSlow write method, "outb" only needs to be written as "outb %1" and not "outb %1 %0".
When %0 is included %ax is generated as junk.
The compiler must include both operands automatically or the processor implicitly uses the A registry and only requires the address.
It is more likely that the compiler is solving it considering that both the x86 and x64 documentation state both operands for the instruction.
There's no clear documentation for the specific "outb" instruction but there is for the "out" instruction.
I left the Port8Bit write method intenionally unchanged to prove that the compiler is skipping it because the inilne gets optimized by the compiler and is never called.
When using inline is that the method needs to be defined inside the headers to have access to the object fields. Maybe I can correct that in the future.

Deleted old portb files.